refactor: import csv file working
