+++ src/Util/TimezoneUtilAwareInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of cgoit\contao-calendar-ical-php8-bundle for Contao Open Source CMS.
+ *
+ * @copyright  Copyright (c) 2023, cgoIT
+ * @author     cgoIT <https://cgo-it.de>
+ * @license    LGPL-3.0-or-later
+ */
+
+namespace Cgoit\ContaoCalendarIcalBundle\Util;
+
+interface TimezoneUtilAwareInterface
+{
+    public function setTimezoneUtil(TimezoneUtil $timezoneUtil): void;
+}
